+++ flask_restplus/swagger.py
@@ -132,6 +132,14 @@ def parse_docstring(obj):
     return parsed
 
 
+def is_hidden(resource):
+    '''
+    Determine whether a Resource has been hidden from Swagger documentation
+    i.e. by using Api.doc(False) decorator
+    '''
+    return hasattr(resource, "__apidoc__") and resource.__apidoc__ is False
+
+
 class Swagger(object):
     '''
     A Swagger documentation wrapper for an API instance.
@@ -224,7 +232,12 @@ class Swagger(object):
             tags.append(tag)
             by_name[tag['name']] = tag
         for ns in api.namespaces:
+            # hide namespaces without any Resources
             if not ns.resources:
+                continue
+            # hide namespaces with all Resources hidden from Swagger documentation
+            resources = (resource for resource, urls, kwargs in ns.resources)
+            if all(is_hidden(r) for r in resources):
                 continue
             if ns.name not in by_name:
                 tags.append({